backend/requirements.txt
# 🚀 KR-AI-Engine - Minimal Requirements
# Nur die wichtigsten Abhängigkeiten für den Start

# ================================
# CORE API FRAMEWORK
# ================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# ================================
# DATABASE # DATABASE & SUPABASE OBJECT STORAGE
# ================================
supabase>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# ================================
# DOCUMENT PROCESSING
# ================================
PyMuPDF>=1.23.0
python-docx>=0.8.0
pytesseract>=0.3.0
Pillow>=10.0.0
pdfplumber>=0.9.0
python-magic>=0.4.0

# ================================
# AI & MACHINE LEARNING
# ================================
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ================================
# HTTP & API CLIENTS
# ================================
httpx>=0.26.0
aiohttp>=3.8.0
ollama>=0.1.0

# ================================
# CACHING & STORAGE
# ================================
redis>=5.0.0

# ================================
# TEXT PROCESSING
# ================================
nltk>=3.8.0
regex>=2023.10.0
unidecode>=1.3.0

# ================================
# UTILITIES
# ================================
tqdm>=4.66.0
rich>=13.6.0
loguru>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.0
click>=8.1.0
python-dateutil>=2.8.0

# ================================
# SECURITY
# ================================
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
